Implemented selecting a photo for query through the built-in Android INTENT_PICK. Began implementing an interface which allows the user to select particular albums to be included in the indexing. The results are divided into local and server and implemented as a SearchResultFragment holding a RecyclerView within a SearchActivity.
Fixed a bug in ColorDescriptor.cpp which results in an exception when trying to read an invalid image (e.g. of size 0 bytes) and included a code snippet to easily calculate the chi-squared distance between color histograms. Began implementing the ability to compare against the database; very promising results; almost instantaneous with around 500 photos.